Moved toolbar-part from inspection to seperate component, `MetadataToolbar.jsx`
